tests/User/requirements.txt
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
asyncpg>=0.29.0
PyYAML>=6.0

